ENHANCE: (breaking change) Make Matrix opaque
I've changed the Matrix type to an opaque one which ensures all the lines
have the same length.

The new type is:
type Matrix a =
    Matrix Width (Array (Array a))

where Width a simple alias to Int.

I didn't put the height in the constructor since it is easy to get it throught
Array.length.

This commit does'nt change the implementation otherwise, it only deals with
(un)wrap the Matrix constructor.